ISSUE #5545 - Groups gets updated when the property changes value
